+++ hw2000/I_MCW.java
@@ -7,7 +7,7 @@ public class I_MCW implements Instruction {
 			a = sys.readMem(sys.AAR);
 			sys.incrAAR(-1);
 			b = sys.readMem(sys.BAR);
-			b = (b & 0100) | (a & ~0100);
+			b = (byte)((b & 0100) | (a & ~0100));
 			sys.writeMem(sys.BAR, b);
 			sys.incrBAR(-1);
 		} while ((a & 0100) == 0 && (b & 0100) == 0);
