+++ hw2000/I_SIB.java
@@ -3,13 +3,13 @@ public class I_SIB implements Instruction {
 	public void execute(HW2000 sys) {
 		byte a;
 		a = sys.IBR;
-		sys.writeChar(sys.AAR, (a >> 6));
+		sys.writeChar(sys.AAR, (byte)(a >> 6));
 		sys.incrAAR(-1);
 		sys.writeChar(sys.AAR, a);
 		sys.incrAAR(-1);
 		if (sys.hadB()) {
 			a = sys.BRR;
-			sys.writeChar(sys.BAR, (a >> 6));
+			sys.writeChar(sys.BAR, (byte)(a >> 6));
 			sys.incrBAR(-1);
 			sys.writeChar(sys.BAR, a);
 			sys.incrBAR(-1);
